requirements.txt
python-telegram-bot==22.0
gspread==5.12.4
oauth2client==4.1.3
python-dotenv==1.0.1
gunicorn==21.2.0
